+++ pitch up.qml
@@ -16,7 +16,7 @@ MuseScore {
         }
       }
 
-      version:  "2.1.1"
+      version:  "2.1.2"
       description: "Raises selection (Shift-click) or individually selected notes (Ctrl-click) by 1 step of n EDO."
       menuPath: "Plugins.n-EDO.Raise Pitch By 1 Step"